Primer refactor de sugerencias
